sheep/ops: fix potential type conversion error
Similar to lib/logger, epoch_log_read_with_timestamp will return -1
for any failure and it cannot be assigned to a variable in type of
size_t. Otherwise, the error branch will never be reached.

+++ sheep/ops.c
@@ -449,7 +449,7 @@ static int local_stat_cluster(struct request *req)
 	max_elogs = req->rq.data_length / sizeof(*elog);
 	epoch = get_latest_epoch();
 	for (i = 0; i < max_elogs; i++) {
-		size_t nr_nodes;
+		int nr_nodes;
 
 		if (epoch <= 0)
 			break;
